added moving functions for cars, and connections between Nodes
Cars now get the path from their current Node to the next one on their
way.
The complete Path has to be given to the car as an array at the
beginning. It then itterates the nodes, with a set movement-speed
definded by the distance between x1y1 & x2y2

also slight fixes in the bigCity & smallCity class